+++ src/client/types/transaction.ts
@@ -1,0 +1,10 @@
+export type TxId = {
+  lt: number;
+  hash: string;
+  contract_address?: string;
+};
+
+export type WalletTxIdResponse = {
+  tx_id?: TxId;
+  wallet_seqno?: number;
+};